arm64 : Remove awkward jumptable hack
We now store return addresses / link register on the stack when we
perform call/returns.

Caveat: Actual addresses on the stack are different from the ones
emitted by bme. That way, modifying the return address on the stack
directly (which should be considered undefined behaviour anyways) will
act differently in native codegen than in the emulation.

When we perform a call, we store the current link register (address of
the callee) to the stack and branch. When returning, we store the lr
in x9 temporarily, pop out into the link register and jump to the
value in x9.

Minor performance increase as expected due to less emitted instructions.